Move all the preference pages to PreferencePages subfolder.
