+++ src/Gui/PreferencePages/DlgSettingsEditor.ui
