+++ src/Gui/PreferencePages/DlgSettingsNotificationArea.cpp
