+++ src/Gui/PreferencePages/DlgSettingsNotificationArea.ui
