+++ src/Gui/PreferencePages/DlgSettingsPythonConsole.cpp
