+++ src/Gui/PreferencePages/DlgSettingsPythonConsole.ui
