+++ src/Gui/PreferencePages/DlgSettingsReportView.cpp
