+++ src/Gui/PreferencePages/DlgSettingsReportView.h
@@ -21,58 +21,39 @@
  ***************************************************************************/
 
 
-#ifndef GUI_DIALOG_DLGSETTINGSEDITOR_H
-#define GUI_DIALOG_DLGSETTINGSEDITOR_H
+#ifndef GUI_DIALOG_DLG_SETTINGS_REPORT_VIEW
+#define GUI_DIALOG_DLG_SETTINGS_REPORT_VIEW
 
-#include "PropertyPage.h"
+#include <Gui/PropertyPage.h>
 #include <memory>
 
-class QTreeWidgetItem;
-
 namespace Gui {
-class PythonSyntaxHighlighter;
-
 namespace Dialog {
-class Ui_DlgSettingsEditor;
+class Ui_DlgSettingsReportView;
 
-/** This class implements a preferences page for the editor settings.
- *  Here you can change different color settings and font for editors.
- *  @author Werner Mayer
+/** The DlgSettingsReportView class implements the available settings for the
+ * Report View to change.
+ * \author Werner Mayer
  */
-struct DlgSettingsEditorP;
-class DlgSettingsEditorImp : public PreferencePage
+class DlgSettingsReportView : public PreferencePage
 {
     Q_OBJECT
 
 public:
-    explicit DlgSettingsEditorImp( QWidget* parent = nullptr );
-    ~DlgSettingsEditorImp() override;
+    explicit DlgSettingsReportView( QWidget* parent = nullptr );
+    ~DlgSettingsReportView() override;
 
-public:
     void saveSettings() override;
     void loadSettings() override;
 
-private:
-    void setupConnections();
-    void onDisplayItemsCurrentItemChanged(QTreeWidgetItem *i);
-    void onColorButtonChanged();
-    void onFontFamilyActivated(const QString&);
-    void onFontSizeValueChanged(const QString&);
-
 protected:
     void changeEvent(QEvent *e) override;
-    void setEditorTabWidth(int);
 
 private:
-    std::unique_ptr<Ui_DlgSettingsEditor> ui;
-    DlgSettingsEditorP* d;
-    Gui::PythonSyntaxHighlighter* pythonSyntax;
-
-    DlgSettingsEditorImp( const DlgSettingsEditorImp & );
-    DlgSettingsEditorImp& operator=( const DlgSettingsEditorImp & );
+    std::unique_ptr<Ui_DlgSettingsReportView> ui;
 };
 
 } // namespace Dialog
 } // namespace Gui
 
-#endif // GUI_DIALOG_DLGSETTINGSEDITOR_H
+#endif //GUI_DIALOG_DLG_SETTINGS_REPORT_VIEW
