+++ src/Gui/PreferencePages/DlgSettingsReportView.ui
