+++ src/Gui/PreferencePages/DlgSettingsViewColor.ui
